requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
pytesseract==0.3.10
opencv-python==4.10.0.84
Pillow==10.4.0
pdf2image==1.16.3
numpy==2.1.1
